data-factory/requirements.txt
feedparser==6.0.11
web3==6.15.1
requests==2.31.0
beautifulsoup4==4.12.3
yfinance==0.2.36
schedule==1.2.1
python-dotenv==1.0.1
openai==1.12.0
lxml==5.1.0

